feat: Complete rewrite of Node.js application to Python/Flask
This commit completes the full rewrite of the original Node.js/Express application into a new, functional Python/Flask application.

The new application includes the following features:
- A complete Flask project structure with models, forms, routes, and templates, cleaned of all legacy Node.js files.
- A full user authentication system (register, login, logout) using Flask-Login.
- Complete CRUD functionality for "protocolos".
- A system for file attachments, storing files directly in the PostgreSQL database.
- A dynamic, filterable dashboard with charts powered by Chart.js and a custom API endpoint.
- Interactive modals for updating and forwarding protocols.
- A feature to export filtered protocol data to an Excel (.xlsx) file.
- A database initialization script (`init_db.py`).
- A `nixpacks.toml` file to ensure correct deployment on Railway.
- All sensitive information (SECRET_KEY, DATABASE_URL) is loaded from environment variables.
