--- iniciar-servidor.bat
@@ -1,5 +1,0 @@
-@echo off
-cd /d "%~dp0"
-echo Iniciando o servidor do Sistema de Protocolo...
-node server.js
-pause
