--- package.json
@@ -1,28 +1,0 @@
-{
-  "name": "protocolo-backend",
-  "version": "1.0.0",
-  "description": "Sistema de Protocolo",
-  "main": "server.js",
-  "scripts": {
-    "start": "node server.js",
-    "test": "dotenv -e .env.test -- jest"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@supabase/supabase-js": "^2.43.4",
-    "cors": "^2.8.5",
-    "exceljs": "^4.4.0",
-    "express": "^4.19.2",
-    "multer": "^1.4.5-lts.1",
-    "pg": "^8.11.3",
-    "dotenv": "^17.2.1",
-    "bcrypt": "^6.0.0",
-    "jsonwebtoken": "^9.0.2"
-  },
-  "devDependencies": {
-    "jest": "^30.1.1",
-    "supertest": "^7.1.4",
-    "dotenv-cli": "^7.4.2"
-  }
-}
