--- supabase-client.js
@@ -1,1 +1,0 @@
-// supabase-client
